% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyTZ.R
\name{useShinyTZ}
\alias{useShinyTZ}
\title{Include shinyTZ JavaScript in Shiny UI}
\usage{
useShinyTZ()
}
\value{
An HTML tag object that should be included in the Shiny UI
}
\description{
This function must be called from within the UI of a Shiny app to enable
automatic timezone detection. It loads the JavaScript code that detects
the browser's timezone and sends it to the server.
}
\examples{
\dontrun{
library(shiny)
library(shinyTZ)

ui <- fluidPage(
  useShinyTZ(),  # Enable timezone detection
  
  h3("Current Time:"),
  datetimeOutput("current_time")
)

server <- function(input, output, session) {
  output$current_time <- renderDatetime({
    Sys.time()
  })
}

shinyApp(ui, server)
}
}
