% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{shinyTZ-package}
\alias{shinyTZ}
\alias{shinyTZ-package}
\title{shinyTZ: Timezone-Aware Date and Time Rendering for Shiny}
\description{
Provides drop-in replacements for Shiny's text outputs that automatically handle timezone conversion based on the user's browser timezone. Eliminates the common problem of server timestamps being displayed in the wrong timezone for remote users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gwarnes-mdsol/shinyTZ}
  \item Report bugs at \url{https://github.com/gwarnes-mdsol/shinyTZ/issues}
}

}
\author{
\strong{Maintainer}: Gregory Warnes \email{warnes@binghamton.edu}

}
\keyword{internal}
